gateway/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==1.10.13
pyyaml==6.0.2
websockets==13.1
